Began reorganizing the settings applications and scaffolding the World Settings application.
