+++ src/applications/settings/parts/CheckboxSetting.svelte
