+++ src/applications/settings/parts/ExhaustionSetting.svelte
