+++ src/applications/settings/parts/NumberInputSetting.svelte
