+++ src/applications/settings/parts/SelectSetting.svelte
