+++ src/applications/settings/tabs/FeaturesSettingsTab.svelte
