+++ src/applications/settings/tabs/InfoTab.svelte
