+++ src/applications/settings/tabs/NpcSettingsTab.svelte
@@ -1,11 +1,11 @@
 <script lang="ts">
   import { FoundryAdapter } from 'src/foundry/foundry-adapter';
-  import CheckboxSetting from 'src/applications/sheet-settings/parts/CheckboxSetting.svelte';
-  import NumberInputSetting from 'src/applications/sheet-settings/parts/NumberInputSetting.svelte';
+  import CheckboxSetting from 'src/applications/settings/parts/CheckboxSetting.svelte';
+  import NumberInputSetting from 'src/applications/settings/parts/NumberInputSetting.svelte';
   import { getContext } from 'svelte';
   import type { Writable } from 'svelte/store';
   import { SettingsProvider } from 'src/settings/settings';
-  import SelectSetting from 'src/applications/sheet-settings/parts/SelectSetting.svelte';
+  import SelectSetting from 'src/applications/settings/parts/SelectSetting.svelte';
   import type {
     SettingsSheetContext,
     SettingsSheetFunctions,
