+++ src/applications/settings/tabs/PlayerSettingsTab.svelte
@@ -3,9 +3,9 @@
   import { SettingsProvider } from 'src/settings/settings';
   import { getContext } from 'svelte';
   import type { Writable } from 'svelte/store';
-  import CheckboxSetting from 'src/applications/sheet-settings/parts/CheckboxSetting.svelte';
-  import NumberInputSetting from 'src/applications/sheet-settings/parts/NumberInputSetting.svelte';
-  import TextInputSetting from 'src/applications/sheet-settings/parts/TextInputSetting.svelte';
+  import CheckboxSetting from 'src/applications/settings/parts/CheckboxSetting.svelte';
+  import NumberInputSetting from 'src/applications/settings/parts/NumberInputSetting.svelte';
+  import TextInputSetting from 'src/applications/settings/parts/TextInputSetting.svelte';
   import SelectSetting from '../parts/SelectSetting.svelte';
   import type {
     SettingsSheetContext,
