+++ src/applications/settings/tabs/VehicleSettingsTab.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
   import { FoundryAdapter } from 'src/foundry/foundry-adapter';
-  import CheckboxSetting from 'src/applications/sheet-settings/parts/CheckboxSetting.svelte';
+  import CheckboxSetting from 'src/applications/settings/parts/CheckboxSetting.svelte';
   import { getContext } from 'svelte';
   import type { Writable } from 'svelte/store';
   import { SettingsProvider } from 'src/settings/settings';
-  import NumberInputSetting from 'src/applications/sheet-settings/parts/NumberInputSetting.svelte';
+  import NumberInputSetting from 'src/applications/settings/parts/NumberInputSetting.svelte';
   import SelectSetting from '../parts/SelectSetting.svelte';
   import type {
     SettingsSheetContext,
