+++ src/applications/world-settings/WorldSettings.svelte
@@ -1,0 +1,9 @@
+<script lang="ts">
+    import type { Tab } from "src/types/types";
+
+    let selectedTabId: string;
+
+    let tabs: Tab[] = [];
+
+    
+</script>
